backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
xgboost
requests
twilio
APScheduler
joblib
python-dotenv